requirements.txt
streamlit>=1.35.0
Pillow>=10.0.0
streamlit-analytics2>=0.5.0
